Add separate .css file for note-related text and package note classes
