+++ src/main/java/seedu/address/logic/commands/notecommands/AddWeaknessCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.notecommands;
 
 import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
 import static seedu.address.model.Model.PREDICATE_SHOW_ALL_PERSONS;
@@ -8,45 +8,47 @@ import java.util.List;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.note.Note;
 import seedu.address.model.person.Person;
 
 /**
- * Delete a weakness from a person in the address book
+ * Adds a weakness note to a person in the address book.
  */
-public class DeleteWeaknessCommand extends Command {
-    public static final String COMMAND_WORD = "weakness-del";
+public class AddWeaknessCommand extends Command {
+    public static final String COMMAND_WORD = "weakness-add";
 
     public static final String MESSAGE_USAGE = COMMAND_WORD
-            + ": Delete a weakness from the weakness-list of a selected person from our contact list. "
+            + ": Add a weakness to a selected person from our contact list. "
             + "Parameters: "
             + "INDEX (must be a positive integer) "
-            + "NOTE-INDEX (must be a positive)\n"
+            + "NOTE_DESCRIPTION (must be non-empty and not more than 50 characters)\n"
             + "Example: " + COMMAND_WORD + " "
             + "1 "
-            + "2";
+            + "Poor endurance";
 
     public static final String MESSAGE_NOT_IMPLEMENTED_YET =
-            "DeleteWeakness command not implemented yet";
+            "AddWeakness command not implemented yet";
 
-    public static final String MESSAGE_SUCCESS = "Weakness has been deleted: %1$s";
+    public static final String MESSAGE_SUCCESS = "New weakness added: %1$s";
 
-    public static final String MESSAGE_ARGUMENTS = "Index: %1$d, Index: %2$d";
+    public static final String MESSAGE_ARGUMENTS = "Index: %1$d, Weakness: %2$s";
 
     private final Index index;
-    private final Index noteIndex;
+    private final Note weakness;
 
     /**
-     * Constructor of DeleteWeaknessCommand class
-     * @param index index of the person in the filtered person list
-     * @param noteIndex index of the weakness from the person's weakness-list to be deleted
+     * @param index of the person in the filtered person list to add the weakness
+     * @param weakness of the person to be added
      */
-    public DeleteWeaknessCommand(Index index, Index noteIndex) {
-        requireAllNonNull(index, noteIndex);
+    public AddWeaknessCommand(Index index, Note weakness) {
+        requireAllNonNull(index, weakness);
+
         this.index = index;
-        this.noteIndex = noteIndex;
+        this.weakness = weakness;
     }
 
     @Override
@@ -58,17 +60,13 @@ public class DeleteWeaknessCommand extends Command {
         }
 
         Person personToEdit = lastShownList.get(index.getZeroBased());
-        List<Note> newWeaknesses = new ArrayList<>(personToEdit.getWeaknesses());
-
-        if (noteIndex.getZeroBased() >= newWeaknesses.size()) {
-            throw new CommandException(Messages.MESSAGE_INVALID_NOTE_DISPLAYED_INDEX);
-        }
-        newWeaknesses.remove(noteIndex.getZeroBased());
+        List<Note> newWeakness = new ArrayList<>(personToEdit.getWeaknesses());
+        newWeakness.add(weakness);
 
         Person editedPerson = new Person(
                 personToEdit.getName(), personToEdit.getPhone(), personToEdit.getEmail(),
-                personToEdit.getAddress(), personToEdit.getTags(),
-                personToEdit.getStrengths(), newWeaknesses, personToEdit.getMiscellaneous());
+                personToEdit.getAddress(), personToEdit.getTags(), personToEdit.getStrengths(), newWeakness,
+                personToEdit.getMiscellaneous());
 
         model.setPerson(personToEdit, editedPerson);
         model.updateFilteredPersonList(PREDICATE_SHOW_ALL_PERSONS);
@@ -86,9 +84,9 @@ public class DeleteWeaknessCommand extends Command {
 
     @Override
     public boolean equals(Object other) {
-        return other == this
-                || (other instanceof DeleteWeaknessCommand
-                && index.equals(((DeleteWeaknessCommand) other).index)
-                && noteIndex.equals(((DeleteWeaknessCommand) other).noteIndex));
+        return other == this // short circuit if same object
+                || (other instanceof AddWeaknessCommand // instanceof handles nulls
+                && (index.equals(((AddWeaknessCommand) other).index)
+                    && weakness.equals(((AddWeaknessCommand) other).weakness)));
     }
 }
