+++ src/main/java/seedu/address/logic/commands/notecommands/DeleteMiscCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.notecommands;
 
 import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
 import static seedu.address.model.Model.PREDICATE_SHOW_ALL_PERSONS;
@@ -8,45 +8,47 @@ import java.util.List;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.note.Note;
 import seedu.address.model.person.Person;
 
 /**
- * Adds a miscellaneous note to a person in the address book.
+ * Delete a miscellaneous note from a person in the address book
  */
-public class AddMiscCommand extends Command {
-    public static final String COMMAND_WORD = "misc-add";
+public class DeleteMiscCommand extends Command {
+    public static final String COMMAND_WORD = "misc-del";
 
     public static final String MESSAGE_USAGE = COMMAND_WORD
-            + ": Add a miscellaneous note to a selected person from our contact list. "
+            + ": Delete a miscellaneous note from the note-list of selected person from our contact list. "
             + "Parameters: "
             + "INDEX (must be a positive integer) "
-            + "NOTE_DESCRIPTION (must be non-empty and not more than 50 characters)\n"
+            + "NOTE-INDEX (must be a positive)\n"
             + "Example: " + COMMAND_WORD + " "
             + "1 "
-            + "Surgery Scheduled for tomorrow";
+            + "2";
 
     public static final String MESSAGE_NOT_IMPLEMENTED_YET =
-            "AddMisc command not implemented yet";
+            "DeleteMisc command not implemented yet";
 
-    public static final String MESSAGE_SUCCESS = "New miscellaneous note added: %1$s";
+    public static final String MESSAGE_SUCCESS = "Miscellaneous note has been deleted: %1$s";
 
-    public static final String MESSAGE_ARGUMENTS = "Index: %1$d, Misc: %2$s";
+    public static final String MESSAGE_ARGUMENTS = "Index: %1$d, Index: %2$d";
 
     private final Index index;
-    private final Note misc;
+    private final Index noteIndex;
 
     /**
-     * @param index of the person in the filtered person list to add the misc
-     * @param misc note of the person to be added
+     * Constructor of DeleteMiscCommand class
+     * @param index index of the person in the filtered person list
+     * @param noteIndex index of the misc. note from the person's misc-list to be deleted
      */
-    public AddMiscCommand(Index index, Note misc) {
-        requireAllNonNull(index, misc);
-
+    public DeleteMiscCommand(Index index, Index noteIndex) {
+        requireAllNonNull(index, noteIndex);
         this.index = index;
-        this.misc = misc;
+        this.noteIndex = noteIndex;
     }
 
     @Override
@@ -59,12 +61,16 @@ public class AddMiscCommand extends Command {
 
         Person personToEdit = lastShownList.get(index.getZeroBased());
         List<Note> newMisc = new ArrayList<>(personToEdit.getMiscellaneous());
-        newMisc.add(misc);
+
+        if (noteIndex.getZeroBased() >= newMisc.size()) {
+            throw new CommandException(Messages.MESSAGE_INVALID_NOTE_DISPLAYED_INDEX);
+        }
+        newMisc.remove(noteIndex.getZeroBased());
 
         Person editedPerson = new Person(
                 personToEdit.getName(), personToEdit.getPhone(), personToEdit.getEmail(),
-                personToEdit.getAddress(), personToEdit.getTags(), personToEdit.getStrengths(),
-                personToEdit.getWeaknesses(), newMisc);
+                personToEdit.getAddress(), personToEdit.getTags(),
+                personToEdit.getStrengths(), personToEdit.getWeaknesses(), newMisc);
 
         model.setPerson(personToEdit, editedPerson);
         model.updateFilteredPersonList(PREDICATE_SHOW_ALL_PERSONS);
@@ -82,9 +88,9 @@ public class AddMiscCommand extends Command {
 
     @Override
     public boolean equals(Object other) {
-        return other == this // short circuit if same object
-                || (other instanceof AddMiscCommand // instanceof handles nulls
-                && (index.equals(((AddMiscCommand) other).index)
-                    && misc.equals(((AddMiscCommand) other).misc)));
+        return other == this
+                || (other instanceof DeleteMiscCommand
+                && index.equals(((DeleteMiscCommand) other).index)
+                && noteIndex.equals(((DeleteMiscCommand) other).noteIndex));
     }
 }
