+++ src/test/java/seedu/address/logic/parser/AddressBookParserTest.java
@@ -15,14 +15,14 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.AddCommand;
-import seedu.address.logic.commands.AddMiscCommand;
-import seedu.address.logic.commands.AddStrengthCommand;
-import seedu.address.logic.commands.AddWeaknessCommand;
+import seedu.address.logic.commands.notecommands.AddMiscCommand;
+import seedu.address.logic.commands.notecommands.AddStrengthCommand;
+import seedu.address.logic.commands.notecommands.AddWeaknessCommand;
 import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.DeleteCommand;
-import seedu.address.logic.commands.DeleteMiscCommand;
-import seedu.address.logic.commands.DeleteStrengthCommand;
-import seedu.address.logic.commands.DeleteWeaknessCommand;
+import seedu.address.logic.commands.notecommands.DeleteMiscCommand;
+import seedu.address.logic.commands.notecommands.DeleteStrengthCommand;
+import seedu.address.logic.commands.notecommands.DeleteWeaknessCommand;
 import seedu.address.logic.commands.EditCommand;
 import seedu.address.logic.commands.EditCommand.EditPersonDescriptor;
 import seedu.address.logic.commands.ExitCommand;
